Make the parameters for the st2kv function kwargs
Make the parameters for the st2kv function into kwargs to prepare for adding the parameter for default value. The parameter for default value is optional and would have been assigned None by default. The function would have no way to determine when a key does not exists whether to return a default value of None. By making the parameters into kwargs, the function can see if the default key is in the kwargs to determine what default value to return if a key does not exists. The default value can support returning a value of NoneType without being confused whether a default value is set or not.